refactor(avatar): make avatar-group a separate directory
