+++ src/avatar-group/avatar-group.less
