+++ src/avatar-group/avatar-group.wxml
