+++ src/avatar-group/props.ts
