+++ src/avatar/__test__/index.json
@@ -1,6 +1,6 @@
 {
   "usingComponents": {
-    "t-avatar-group": "../avatar-group",
+    "t-avatar-group": "../../avatar-group/avatar-group",
     "t-avatar": "../avatar"
   }
 }
