+++ src/avatar/_example/exhibition/index.json
@@ -2,6 +2,6 @@
   "component": true,
   "usingComponents": {
     "t-avatar": "tdesign-miniprogram/avatar/avatar",
-    "t-avatar-group": "tdesign-miniprogram/avatar/avatar-group"
+    "t-avatar-group": "tdesign-miniprogram/avatar-group/avatar-group"
   }
 }
